refactor: common function for updating post
